Move the transaction_context class to its own file
