+++ test/cppsuite/CMakeLists.txt
@@ -25,6 +25,7 @@ add_library(cppsuite_test_harness STATIC
     src/main/test.cpp
     src/main/thread_context.cpp
     src/main/throttle.cpp
+    src/main/transaction.cpp
     src/storage/connection_manager.cpp
     src/storage/scoped_connection.cpp
     src/storage/scoped_types.cpp
