+++ test/cppsuite/src/main/thread_context.h
@@ -36,66 +36,12 @@
 #include "src/component/timestamp_manager.h"
 #include "src/main/configuration.h"
 #include "src/storage/scoped_types.h"
+#include "transaction.h"
 
 namespace test_harness {
 enum thread_type { CUSTOM, INSERT, READ, REMOVE, UPDATE };
 
 const std::string type_string(thread_type type);
-
-class transaction_context {
-    public:
-    explicit transaction_context(
-      configuration *config, timestamp_manager *timestamp_manager, WT_SESSION *session);
-
-    bool active() const;
-    void add_op();
-    void begin(const std::string &config = "");
-    /* Begin a transaction if we are not currently in one. */
-    void try_begin(const std::string &config = "");
-    /*
-     * Commit a transaction and return true if the commit was successful.
-     */
-    bool commit(const std::string &config = "");
-    /* Rollback a transaction, failure will abort the test. */
-    void rollback(const std::string &config = "");
-    /* Attempt to rollback the transaction given the requirements are met. */
-    void try_rollback(const std::string &config = "");
-    /* Set a commit timestamp. */
-    int set_commit_timestamp(wt_timestamp_t ts);
-    /* Set that the transaction needs to be rolled back. */
-    void set_needs_rollback(bool rollback);
-    /*
-     * Returns true if a transaction can be committed as determined by the op count and the state of
-     * the transaction.
-     */
-    bool can_commit();
-    /*
-     * Returns true if a transaction can be rolled back as determined by the op count and the state
-     * of the transaction.
-     */
-    bool can_rollback();
-
-    private:
-    bool _in_txn = false;
-    bool _needs_rollback = false;
-
-    /*
-     * _min_op_count and _max_op_count are the minimum and maximum number of operations within one
-     * transaction. is the current maximum number of operations that can be executed in the current
-     * transaction.
-     */
-    int64_t _max_op_count = INT64_MAX;
-    int64_t _min_op_count = 0;
-    /*
-     * op_count is the current number of operations that have been executed in the current
-     * transaction.
-     */
-    int64_t _op_count = 0;
-    int64_t _target_op_count = 0;
-
-    timestamp_manager *_timestamp_manager = nullptr;
-    WT_SESSION *_session = nullptr;
-};
 
 /* Container class for a thread and any data types it may need to interact with the database. */
 class thread_context {
